move unused programs and tests to new folders
