+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/Test_DistanceSensor.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.*;
 import com.qualcomm.robotcore.hardware.DistanceSensor;
