+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/Test_RB1.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+
+import org.firstinspires.ftc.teamcode.CSBase;
 
 /*
  * This OpMode illustrates the concept of driving a path based on encoder counts.
@@ -29,7 +31,7 @@ import com.qualcomm.robotcore.eventloop.opmode.Disabled;
  */
 
 @Disabled
-public class Test_RB3 extends CSBase {
+public class Test_RB1 extends CSBase {
     @Override
     public void runOpMode() {
         setup(true);
@@ -38,7 +40,7 @@ public class Test_RB3 extends CSBase {
         // ------Main Code------
         // ---------------------
 
-        ejectPixel();
+        drive(tilesToInches(-2.1));
 
         telemetry.addData("Path", "Complete");
         telemetry.update();
