+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tests/Test_Strafing20.java
@@ -1,6 +1,8 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.tests;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
+
+import org.firstinspires.ftc.teamcode.CSBase;
 
 @Disabled
 public class Test_Strafing20 extends CSBase {
